menu v3.5 , sistema de logs, eh clean , gestion de archivos mejorados (eliminar,asignar,designar)
