+++ Final/cursosCreados/573490506.txt
@@ -1,0 +1,9 @@
+[Machine Learning Basico]
+id: 573490506
+categoria: Ciencia de Datos
+descripcion: Primeros pasos en modelos predictivos usando Python
+duracion: 60 horas
+fecha: 16/06/2025 19:35:06
+
+[LECCIONES]
+- Regresion lineal
