+++ Final/cursosCreados/7342609.txt
@@ -1,0 +1,9 @@
+[Programacion en python desde cero]
+id: 7342609
+categoria: Programacion Basica
+descripcion: Aprende los fundamentos de python para automatizar tareas
+duracion: 30 horas
+fecha: 16/06/2025 19:32:08
+
+[LECCIONES]
+- Introduccion a Python
