+++ Final/cursosCreados/966782071.txt
@@ -1,0 +1,9 @@
+[Fundamentos de Ciberseguridad]
+id: 966782071
+categoria: Seguridad Informatica
+descripcion: Conceptos muy importantes para proteger sistemas y redes
+duracion: 35 horas
+fecha: 16/06/2025 19:34:10
+
+[LECCIONES]
+- Amenazas comunes
