--- Final/profesoresGuardados/P7.txt
@@ -1,8 +1,0 @@
-codigo: P7
-nombre: AAA
-apellido: AAA
-correo: Aa@gmail.com
-tiempo_en_coursera: 4
-id: 107
-reputacion: 3
-curso_asignado:
